# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "MiniProyecto",
        "version": "1.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api"
        }
    ],
    "paths": {
...
        "/bills?enterpriseId={enterpriseId}&days={days}&clientId={clientId}": {
            "get": {
                "tags": [
                    "Bills"
                ],
                "summary": "Get all bills",
                "operationId": "getBills",
                "parameters": [
                    {
                        "name": "clientId",
                        "in": "query",
                        "description": "First query parameter",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "enterpriseId",
                        "in": "query",
                        "description": "Second query parameter",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "days",
                        "in": "query",
                        "description": "Second query parameter",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Todas las facturas",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "404": {
                        "description": "not found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/bills": {
            "post": {
                "tags": [
                    "Bills"
                ],
                "summary": "Create bill",
                "description": "Create and return one bill",
                "operationId": "create_bill",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "client_id",
                                    "amount",
                                    "item",
                                    "unit_cost"
                                ],
                                "properties": {
                                    "client_id": {
                                        "type": "integer"
                                    },
                                    "amount": {
                                        "type": "number",
                                        "format": "double"
                                    },
                                    "item": {
                                        "format": "string"
                                    },
                                    "unit_cost": {
                                        "type": "number",
                                        "format": "double"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "client_id": 4,
                                    "amount": 124,
                                    "item": "Product descripton",
                                    "unit_cost": 25.54
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Empresa registrada satisfactoriamente",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "404": {
                        "description": "not found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    }
}